deleted animal_database, and created crud file
