+++ animal_forms.php
@@ -1,9 +1,7 @@
-
 <?php
 
-
-require_once 'classes/animal_database.php';
-
+require_once 'classes/dbConnection.php';
+require_once 'classes/crudAnimal.php';
 
 $database = new Database();
