+++ animal_index.php
@@ -1,5 +1,6 @@
 <?php
-require_once 'classes/animal_database.php';
+require_once 'classes/dbConnection.php';
+require_once 'classes/crudAnimal.php';
 
 
 $database = new Database();
